WIP
we have the pit on the screen and we can fall down into the cave.

But we fall in front of the ground. We want to update the pit to be in front of the surface and the ground. And have a small band across the surface to make it look like the player is falling behind it.